+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning v2.0.0](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.2.0] - 2021-03-27
 ### Added
 - Scan error count in the output
 - Reparse point support
@@ -79,4 +79,5 @@ temporary directory.
 [1.0.1]: https://github.com/ForNeVeR/Vacuum/compare/1.0.0...1.0.1
 [1.1.0]: https://github.com/ForNeVeR/Vacuum/compare/1.0.1...v1.1.0
 [1.1.1]: https://github.com/ForNeVeR/Vacuum/compare/v1.1.0...v1.1.1
-[Unreleased]: https://github.com/ForNeVeR/Vacuum/compare/v1.1.1...HEAD
+[1.2.0]: https://github.com/ForNeVeR/Vacuum/compare/v1.1.1...v1.2.0
+[Unreleased]: https://github.com/ForNeVeR/Vacuum/compare/v1.2.0...HEAD